requirements.txt
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
Pillow==8.4.0
plotly==5.4.0
streamlit==0.81.1
